refactor: renomeia pacote do exeption handler
